단위 테스트와 통합 테스트 분리, 각 테스트 코드가 하나의 기능만 테스트하도록 수정
CardGameAppTests 를 DeckTest 와 IntergrationTest 로 분리
shuffle 테스트는 카드를 100번 섞어 한두번쯤 전과 같은 결과가 나오더라도 테스트를 패스하도록 해줌
reset 은 덱 두개를 모두 리셋해준 뒤 두 덱이 같은지 확인해줌
