+++ CardGameApp/DeckTest/Info.plist
